backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
PyJWT
requests
google-generativeai
google-genai
